requirements.txt
tkinter
opencv-python
Pillow
serial
bpy
